// src/enums.ts
export enum Tags {
    DOCUMENT = 'w:document',
    BODY = 'w:body',
    TABLE = 'w:tbl',
    TABLE_ROW = 'w:tr',
    TABLE_CELL = 'w:tc',
    TEXT = 'w:t',
    TEXT_CONTAINER = 'w:r',
    TEXT_CONTAINER_STYLE = 'w:rPr',
    LINE_BREAK = 'w:br',
    INSERT_TEXT = 'w:instrText',
    FIELD_CHAR = 'w:fldChar',
    PROOF = 'w:proofErr',
    BOOKMARK_START = 'w:bookmarkStart',
    BOOKMARK_END = 'w:bookmarkEnd',
    PARAGRAPH = 'w:p'
}